pyproject.toml
[project]
name = "esa-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "slack-bolt>=1.27.0",
]
